requirements.txt
absl-py==2.1.0
alembic==1.14.1
asttokens==3.0.0
attrs==25.2.0
backcall==0.2.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
databricks-sdk==0.49.0
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
docker==7.1.0
executing==2.2.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jax==0.4.13
jaxlib==0.4.13
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
Mako==1.3.9
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mediapipe==0.10.11
ml-dtypes==0.2.0
mlflow==2.17.2
mlflow-skinny==2.17.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opt_einsum==3.4.0
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smmap==5.0.2
sounddevice==0.5.1
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
